refactor: split Mqlv2IrEmitters into Mqlv2ExpressionEmitter + Mqlv2StageEmitter
Decomposes the 1638-line Mqlv2IrEmitters class along its natural
responsibility seam. Mqlv2ExpressionEmitter owns expression/predicate
translation (translateExpression, translatePredicate, all function-call and
predicate arms, all subquery-predicate helpers, inner-pipeline helpers).
Mqlv2StageEmitter owns stage-level translation (translateFromStage through
translateQueryGroupStage). Dependency is strictly one-way: StageEmitter
calls into ExpressionEmitter; ExpressionEmitter does not need StageEmitter.
Shared helpers (simpleColumnName, extractUnnestArrayPath, buildFieldChain,
translateAggregateReference) live in ExpressionEmitter as package-private
statics. Mqlv2SelectTranslator updated to call Mqlv2StageEmitter.*
throughout; Mqlv2IrEmittersTest updated to call Mqlv2ExpressionEmitter.*
for wrapAsSubPipeline tests. Zero behavior change; all tests pass.
